Mounting and Revision
I created a mounting system for the whole project using what will be a lasercut acyrlic baseplate, I modeled the baseplate, a set of L brackets, and a ton of new fasteners. In this hour i also looked over and revised what else was overlooked or would be needed for the project to function, such as bearing retention, mounting holes for an arduino, slots cut for inserting hex nuts for fasteners. I also modeled a set of TPU feet for the project to rest from